Esp32 directory
The directory contains the sketch loaded in the esp32. The directory contains the folder "data" where are located the files that are stored inside the flash memory using SPIFFS library.
These are a webserver folder (contains the files that make up the webserver in Access Point mode for configuration) and a config.txt file (it will contains the wifi parameter and other configuration data). Soon will be uploaded a complete documentation