Database design change
The database design has been modified. The house is identified by an ID and not by the entire address. The new design is more similar to the whatsapp group: every user must to be subscribed to a house and the single user cannot also be subscribed to anything. More informations will be written inside the documentation