Add wifi scan
The code to scan wifi network has been added: when is made a post request to a specified url by the client (using ajax), the esp32 webserver respond by scannig the wifi networks and creating a string with json format (altough the data send is a plain string). The wifi network visualization must be implemented by js on client side. The trasnferring of json string between client and esp32 is perfectly working