Wifi client almost complete
The wifi page has been almost completed. Now there is the functionality to add manually the ssid and password of a wifi network (it still need js validation).
The network scan code for esp32 is completed. Now is needed the functionality that allows esp to connect to wifi network that has been sent by the client and save it inside flash memori (config.txt file using SPIFFS)