Modify of html style
Little updates to the html files. The new credentials.html has been created and it will contain the username and password fields that allows the connections with its own account on the general server.

It still needs to be implemented the errors handling for wrong input data (javascript) and js code need to be refactored