Esp32 - Network: wifi scan, connection, links
The first page of the configuration phase has been fully completed. The esp32 scan the networks, show the result to the user who can also connect to the network. The user will be redirected on the second page only if the esp32 connects to wifi.

The client interact with the esp webserver by continuous and asynchronous js requests (XMLHttpRequest). The esp webserver is asynchronous: the code that scans the wifi networks and connects to them runs in the loop( ) function. When the asynchronous request happen than some global boolean variables are set that allows the code in the loop( ) to run. When the code products a result, the same boolean variables are set differently and the client request can access to the results (it's like the locking and unlocking of a database/operative system resources without actually locking them).

The UI and js code have been improved and the structure of the other html/js files has been created