Raspserver - Update webserver communication
Some links of home.html have been updated with Flask functions url_for( ).

The home.js file has been modified: when the body load a post request is made to the flask server in order to get the json file with all the houses, rooms and devices data. All the objects and the data will be created and shown only when the json has been received (it runs inside the "onload" request call back function).

The request to get the json file of the devices status has not been implemented yet: this file contain the devices status (mqtt implementation required) and will be requested every second but only if someone has interacted with the device